+++ packages/mint-components/src/components/sqm-instant-access-registration/useInstantAccessRegistration.ts
@@ -67,6 +67,12 @@ export function useInstantAccessRegistration(props: InstantAccessRegistration) {
           error: "",
           validationErrors: { email: props.invalidEmailErrorMessage },
         });
+      } else if (error?.name === "fraud_error") {
+        setRegistrationFormState({
+          ...registrationFormState,
+          loading: false,
+          error: "FRAUD",
+        });
       } else {
         setRegistrationFormState({
           ...registrationFormState,