+++ packages/mint-components/src/components/sqm-rewards-table/RewardsTableCell.stories.tsx
@@ -255,6 +255,7 @@ const referral = (user = null) => {
       dateConverted: 1640038417173,
       dateReferralStarted: 1640038417173,
       dateReferralPaid: 1640038417173,
+      dateModerated: 1640038417173,
       dateReferralEnded: null,
       moderationStatus: null,
       rewards: null,
@@ -272,6 +273,7 @@ const referred = (user = null) => {
       dateConverted: 1640038417173,
       dateReferralStarted: 1640038417173,
       dateReferralPaid: 1640038417173,
+      dateModerated: 1640038417173,
       dateReferralEnded: null,
       moderationStatus: null,
       rewards: null,
@@ -395,6 +397,9 @@ export const StatusCellExpired = () => {
 const pending = {
   statuses: ["PENDING"],
 };
+const pendingReview = {
+  statuses: ["PENDING_REVIEW"],
+};
 const us_tax = {
   pendingReasons: ["US_TAX"],
 };
@@ -449,6 +454,64 @@ export const StatusCellPendingUnhandled = () => {
     <sqm-rewards-table-status-cell
       statusText="Pending"
       reward={{ ...rewardsData, ...pending, ...unhandled }}
+    ></sqm-rewards-table-status-cell>
+  );
+};
+
+export const StatusCellPendingReview = () => {
+  return (
+    <sqm-rewards-table-status-cell
+      statusText="Pending"
+      reward={{
+        ...rewardsData,
+        referral: {
+          id: "123",
+          dateConverted: 1640038417173,
+          dateReferralStarted: 1640038417173,
+          dateReferralPaid: 1640038417173,
+          dateModerated: 1640038417173,
+          dateReferralEnded: null,
+          moderationStatus: null,
+          rewards: null,
+          referrerUser: null,
+          referredUser: null,
+          fraudData: {
+            moderationStatus: "PENDING",
+            autoModerationStatus: "PENDING",
+            manualModerationStatus: "PENDING",
+          },
+        },
+      }}
+    ></sqm-rewards-table-status-cell>
+  );
+};
+
+export const StatusCellDenied = () => {
+  return (
+    <sqm-rewards-table-status-cell
+      statusText="Denied"
+      reward={{
+        ...rewardsData,
+        dateCancelled: 10000,
+
+        referral: {
+          id: "123",
+          dateConverted: 1640038417173,
+          dateReferralStarted: 1640038417173,
+          dateReferralPaid: 1640038417173,
+          dateModerated: 1640038417173,
+          dateReferralEnded: null,
+          moderationStatus: null,
+          rewards: null,
+          referrerUser: null,
+          referredUser: null,
+          fraudData: {
+            moderationStatus: "DENIED",
+            autoModerationStatus: "DENIED",
+            manualModerationStatus: "DENIED",
+          },
+        },
+      }}
     ></sqm-rewards-table-status-cell>
   );
 };